requirements.txt
langchain
langchain_openai
langchain_google_genai
langchain_experimental
google-search-results
langchainhub
python-dotenv
vt-py
validators